requirements.txt
flask
scikit-learn
pandas
nltk
google-generativeai
